+++ lib/tri/form/__init__.py
@@ -39,7 +39,7 @@ from tri.declarative import (
     should_show,
     sort_after,
     with_meta,
-)
+    evaluate)
 from tri.struct import Struct
 
 from tri.form.compat import (
@@ -112,33 +112,12 @@ def bool_parse(string_value):
         raise ValueError('%s is not a valid boolean value' % string_value)
 
 
-def foreign_key_factory(model_field, model, **kwargs):
-    del model
-    setdefaults_path(
-        kwargs,
-        choices=model_field.foreign_related_fields[0].model.objects.all(),
-    )
-    return Field.choice_queryset(model_field=model_field, **kwargs)
-
-
 def many_to_many_factory_read_from_instance(field, instance):
     return getattr_path(instance, field.attr).all()
 
 
 def many_to_many_factory_write_to_instance(field, instance, value):
     getattr_path(instance, field.attr).set(value)
-
-
-def many_to_many_factory(model_field, **kwargs):
-    setdefaults_path(
-        kwargs,
-        choices=model_field.remote_field.model.objects.all(),
-        read_from_instance=many_to_many_factory_read_from_instance,
-        write_to_instance=many_to_many_factory_write_to_instance,
-        extra__django_related_field=True,
-    )
-    kwargs['model'] = model_field.remote_field.model
-    return Field.multi_choice_queryset(model_field=model_field, **kwargs)
 
 
 _field_factory_by_field_type = OrderedDict()
@@ -185,10 +164,10 @@ def create_members_from_model(default_factory, model, member_params_by_member_na
             else:
                 members.append(foo)
     assert_kwargs_empty(member_params_by_member_name)
-    return members + (extra if extra is not None else []) + [Field.from_model(model=model, field_name=x) for x in extra_includes]
+    return members + (extra if extra is not None else []) + [default_factory(model=model, field_name=x) for x in extra_includes]
 
 
-def member_from_model(model, factory_lookup, defaults_factory, factory_lookup_register_function=None, field_name=None, model_field=None, **kwargs):
+def member_from_model(cls, model, factory_lookup, defaults_factory, factory_lookup_register_function=None, field_name=None, model_field=None, **kwargs):
     if model_field is None:
         sub_field_name, _, field_path_rest = field_name.partition('__')
 
@@ -197,11 +176,12 @@ def member_from_model(model, factory_lookup, defaults_factory, factory_lookup_re
 
         if field_path_rest:
             result = member_from_model(
+                cls=cls,
                 model=model_field.remote_field.model,
                 factory_lookup=factory_lookup,
                 defaults_factory=defaults_factory,
                 factory_lookup_register_function=factory_lookup_register_function,
-                field_name=sub_field_name,
+                field_name=field_path_rest,
                 **kwargs)
             result.name = field_name
             result.attr = field_name
@@ -211,14 +191,15 @@ def member_from_model(model, factory_lookup, defaults_factory, factory_lookup_re
         kwargs,
         defaults_factory(model_field),
         name=field_name,
+        call_target__cls=cls,
     )
 
     factory = factory_lookup.get(type(model_field), MISSING)
 
     if factory is MISSING:
-        for django_field_type, func in reversed(factory_lookup.items()):
+        for django_field_type, foo in reversed(factory_lookup.items()):
             if isinstance(model_field, django_field_type):
-                factory = func
+                factory = foo
                 break  # pragma: no mutate optimization
 
     if factory is MISSING:
@@ -227,25 +208,32 @@ def member_from_model(model, factory_lookup, defaults_factory, factory_lookup_re
             message += ' Register a factory with %s, you can also register one that returns None to not handle this field type' % factory_lookup_register_function.__name__
         raise AssertionError(message)
 
-    return factory(model_field=model_field, model=model, **kwargs) if factory else None
+    if factory:
+        factory = evaluate(factory, model_field=model_field, field_name=field_name)
+    return factory(model_field=model_field, model=model, **kwargs) if factory is not None else None
 
 
 @dispatch(
     field=EMPTY,
 )
-def expand_member(model, factory_lookup, defaults_factory, name, field, field_name=None, model_field=None):
+def expand_member(cls, model, factory_lookup, defaults_factory, name, field, field_name=None, model_field=None):
     if model_field is None:
         # noinspection PyProtectedMember
         model_field = model._meta.get_field(field_name)
 
-    result = [member_from_model(model=model_field.remote_field.model,
-                                factory_lookup=factory_lookup,
-                                factory_lookup_register_function=register_field_factory,
-                                defaults_factory=defaults_factory,
-                                field_name=sub_model_field.name,
-                                name=name + '__' + sub_model_field.name,
-                                **field.pop(sub_model_field.name, {}))
-              for sub_model_field in get_fields(model=model_field.remote_field.model)]
+    result = [
+        member_from_model(
+            cls=cls,
+            model=model_field.remote_field.model,
+            factory_lookup=factory_lookup,
+            factory_lookup_register_function=register_field_factory,
+            defaults_factory=defaults_factory,
+            field_name=sub_model_field.name,
+            name=name + '__' + sub_model_field.name,
+            **field.pop(sub_model_field.name, {})
+        )
+        for sub_model_field in get_fields(model=model_field.remote_field.model)
+    ]
     assert_kwargs_empty(field)
     return [x for x in result if x is not None]
 
@@ -351,13 +339,11 @@ datetime_iso_formats = [
 
 
 def datetime_parse(string_value, **_):
-    errors = []
     for iso_format in datetime_iso_formats:
         try:
             return datetime.strptime(string_value, iso_format)
-        except ValueError as e:
-            errors.append('%s' % e)
-    assert errors
+        except ValueError:
+            pass
     raise ValidationError('Time data "%s" does not match any of the formats %s' % (string_value, ', '.join('"%s"' % x for x in datetime_iso_formats)))
 
 
@@ -850,9 +836,10 @@ class Field(RefinableObject):
     def __repr__(self):
         return '<{}.{} {}>'.format(self.__class__.__module__, self.__class__.__name__, self.name)
 
-    @staticmethod
-    def from_model(model, field_name=None, model_field=None, **kwargs):
+    @classmethod
+    def from_model(cls, model, field_name=None, model_field=None, **kwargs):
         return member_from_model(
+            cls=cls,
             model=model,
             factory_lookup=_field_factory_by_field_type,
             factory_lookup_register_function=register_field_factory,
@@ -861,9 +848,10 @@ class Field(RefinableObject):
             model_field=model_field,
             **kwargs)
 
-    @staticmethod
-    def from_model_expand(model, field_name=None, model_field=None, **kwargs):
+    @classmethod
+    def from_model_expand(cls, model, field_name=None, model_field=None, **kwargs):
         return expand_member(
+            cls=cls,
             model=model,
             factory_lookup=_field_factory_by_field_type,
             defaults_factory=field_defaults_factory,
@@ -973,7 +961,7 @@ class Field(RefinableObject):
 
     @classmethod
     @class_shortcut(
-        class_call_target="choice",
+        call_target__attribute="choice",
         choices=[True, False],
         parse=choice_parse,
         choice_to_option=lambda form, field, choice, **_: (
@@ -989,7 +977,7 @@ class Field(RefinableObject):
 
     @classmethod
     @class_shortcut(
-        class_call_target="choice",
+        call_target__attribute="choice",
         parse=choice_queryset_parse,
         choice_to_option=choice_queryset_choice_to_option,
         endpoint_path=choice_queryset_endpoint_path,
@@ -1017,7 +1005,7 @@ class Field(RefinableObject):
 
     @classmethod
     @class_shortcut(
-        class_call_target="choice",
+        call_target__attribute="choice",
         attrs__multiple=True,
         choice_to_option=multi_choice_choice_to_option,
         is_list=True,
@@ -1027,7 +1015,7 @@ class Field(RefinableObject):
 
     @classmethod
     @class_shortcut(
-        class_call_target="choice_queryset",
+        call_target__attribute="choice_queryset",
         attrs__multiple=True,
         choice_to_option=multi_choice_queryset_choice_to_option,
         is_list=True,
@@ -1037,7 +1025,7 @@ class Field(RefinableObject):
 
     @classmethod
     @class_shortcut(
-        class_call_target="choice",
+        call_target__attribute="choice",
         input_template='tri_form/radio.html',
     )
     def radio(cls, call_target=None, **kwargs):
@@ -1135,51 +1123,32 @@ class Field(RefinableObject):
     def phone_number(cls, call_target=None, **kwargs):
         return call_target(**kwargs)
 
-    @staticmethod
-    @shortcut
-    @dispatch
-    def comma_separated(parent_field):
-        """
-        Shortcut to create a comma separated list of something. You can use this to create a comma separated text input that gives nice validation errors easily. Example:
+    @classmethod
+    @class_shortcut(
+        call_target__attribute='choice_queryset',
+    )
+    def foreign_key(cls, model_field, model, call_target, **kwargs):
+        del model
+        setdefaults_path(
+            kwargs,
+            choices=model_field.foreign_related_fields[0].model.objects.all(),
+        )
+        return call_target(model_field=model_field, **kwargs)
 
-        .. code:: python
-
-            Field.comma_separated(Field.email)
-
-        :type parent_field: Field
-        """
-        new_field = copy.copy(parent_field)
-
-        def parse_comma_separated(form, field, string_value):
-            errors = []
-            result = []
-            for x in string_value.split(','):
-                x = x.strip()
-                try:
-                    result.append(parent_field.parse(form=form, field=field, string_value=x.strip()))
-                except ValueError as e:
-                    errors.append('Invalid value "%s": %s' % (x, e))
-                except ValidationError as e:
-                    for message in e.messages:
-                        errors.append('Invalid value "%s": %s' % (x, message))
-            if errors:
-                raise ValidationError(errors)
-            return ', '.join(result)
-
-        new_field.parse = parse_comma_separated
-
-        def is_valid_comma_separated(form, field, parsed_data):
-            errors = set()
-            for x in parsed_data.split(','):
-                x = x.strip()
-                is_valid, error = parent_field.is_valid(form=form, field=field, parsed_data=x)
-                if not is_valid:
-                    errors.add('Invalid value "%s": %s' % (x, error))
-            return errors == set(), errors
-
-        new_field.is_valid = is_valid_comma_separated
-
-        return new_field
+    @classmethod
+    @class_shortcut(
+        call_target__attribute='multi_choice_queryset',
+    )
+    def many_to_many(cls, call_target, model_field, **kwargs):
+        setdefaults_path(
+            kwargs,
+            choices=model_field.remote_field.model.objects.all(),
+            read_from_instance=many_to_many_factory_read_from_instance,
+            write_to_instance=many_to_many_factory_write_to_instance,
+            extra__django_related_field=True,
+        )
+        kwargs['model'] = model_field.remote_field.model
+        return call_target(model_field=model_field, **kwargs)
 
 
 def get_fields(model):
@@ -1235,11 +1204,11 @@ class Form(RefinableObject):
     """ :type: tri.declarative.Namespace """
     base_template = Refinable()
     """ :type: str """
-    field_class = Refinable()
+    member_class = Refinable()
 
     class Meta:
         base_template = 'tri_form/base.html'
-        field_class = Field
+        member_class = Field
 
     @dispatch(
         is_full_form=True,
@@ -1370,7 +1339,7 @@ class Form(RefinableObject):
     def fields_from_model(cls, field, **kwargs):
         return create_members_from_model(
             member_params_by_member_name=field,
-            default_factory=cls.get_meta().field_class.from_model,
+            default_factory=cls.get_meta().member_class.from_model,
             **kwargs
         )